Transaction added

+++ app/src/main/java/org/pesmypetcare/mypetcare/features/users/User.java
@@ -269,7 +269,7 @@ public class User {
         return achievements;
     }
 
-    public void updateAchievementProgress(String name, Double progress) throws InvalidFormatException {
+    public void updateAchievementProgress(String name, Integer progress) throws InvalidFormatException {
         for (UserAchievement userAchievement : achievements) {
             if(userAchievement.getName().equals(name)) {
                 userAchievement.updateProgress(progress);
@@ -298,6 +298,7 @@ public class User {
         levels.add(1, 50);
         commentator.setLevels(levels);
     }
+
     private void initializeContributor() {
         UserAchievement contributor = new UserAchievement("Contributor", R.string.contributor_desc);
         List<Integer> levels = new ArrayList<>();
@@ -307,6 +308,7 @@ public class User {
         levels.add(1, 40);
         contributor.setLevels(levels);
     }
+
     private void initializeWalker() {
         UserAchievement walker = new UserAchievement("Walker", R.string.walker_desc);
         List<Integer> levels = new ArrayList<>();
@@ -316,6 +318,7 @@ public class User {
         levels.add(1, 200);
         walker.setLevels(levels);
     }
+
     private void initializeZoo() {
         UserAchievement zoo = new UserAchievement("Zoo", R.string.walker_desc);
         List<Integer> levels = new ArrayList<>();
@@ -325,6 +328,7 @@ public class User {
         levels.add(1, 4);
         zoo.setLevels(levels);
     }
+
     private void initializeSuperwalker() {
         UserAchievement superwalker = new UserAchievement("Superwalker", R.string.superwalker_desc);
         List<Integer> levels = new ArrayList<>();
@@ -334,6 +338,7 @@ public class User {
         levels.add(1, 100);
         superwalker.setLevels(levels);
     }
+
     private void initializeCleaner() {
         UserAchievement cleaner = new UserAchievement("Clean as a whistle", R.string.clean_desc);
         List<Integer> levels = new ArrayList<>();
@@ -343,6 +348,7 @@ public class User {
         levels.add(1, 20);
         cleaner.setLevels(levels);
     }
+
     private void initializeScaler() {
         UserAchievement scaler = new UserAchievement("Master of scale", R.string.scale_desc);
         List<Integer> levels = new ArrayList<>();
@@ -352,6 +358,7 @@ public class User {
         levels.add(1, 20);
         scaler.setLevels(levels);
     }
+
     private void initializeFounder() {
         UserAchievement founder = new UserAchievement("Founder", R.string.founder_desc);
         List<Integer> levels = new ArrayList<>();
@@ -361,6 +368,7 @@ public class User {
         levels.add(1, 10);
         founder.setLevels(levels);
     }
+
     private void iniitializeMealer() {
         UserAchievement mealer = new UserAchievement("You are what you eat", R.string.meals_desc);
         List<Integer> levels = new ArrayList<>();
@@ -370,6 +378,7 @@ public class User {
         levels.add(1, 200);
         mealer.setLevels(levels);
     }
+
     private void initializeEvent() {
         UserAchievement event = new UserAchievement("All under control", R.string.event_desc);
         List<Integer> levels = new ArrayList<>();
@@ -379,6 +388,5 @@ public class User {
         levels.add(1, 50);
         event.setLevels(levels);
     }
-
 
 }
